+++ Services/DatabaseService.cs
@@ -2,7 +2,6 @@ using System.IO;
 using System.Text;
 using Microsoft.Data.Sqlite;
 using QAMP.Models;
-using QAMP.ViewModels;
 
 namespace QAMP.Services;